+++ types/test/resources.d.ts
@@ -1,0 +1,9 @@
+/**
+ * テストリソース
+ *
+ * @ignore
+ */
+import * as EmailNotificationFactory from '../lib/factory/notification/email';
+export declare namespace notification {
+    function createEmail(): EmailNotificationFactory.INotification;
+}